feat(user): add user management features for locking, banning, and VIP status
Implement backend and frontend changes to allow admins to lock/unlock user accounts, ban/unban users from commenting, and set/remove VIP status. This includes new API endpoints, service methods, and UI updates in the UserManagement component. Also, add checks in AuthService and CommentService to enforce these restrictions.

+++ backend/src/main/java/com/pbl5/pbl5/modal/Chapter.java
@@ -21,7 +21,7 @@ public class Chapter {
     private String title;
     
     @Column(name = "chapter_number")
-    private Integer chapterNumber;
+    private float chapterNumber;
     
     @Column(name = "created_at")
     private LocalDateTime createdAt;
@@ -61,11 +61,11 @@ public class Chapter {
         this.title = title;
     }
 
-    public Integer getChapterNumber() {
+    public float getChapterNumber() {
         return chapterNumber;
     }
 
-    public void setChapterNumber(Integer chapterNumber) {
+    public void setChapterNumber(float chapterNumber) {
         this.chapterNumber = chapterNumber;
     }
